Refactor readers into python module.
